Move zipper inspect into own module and add `default_inspect_as/1`
