+++ lib/sourceror/zipper.ex
@@ -19,70 +19,6 @@ defmodule Sourceror.Zipper do
     * [Zippers](zippers.html) (an introduction)
     * [Expand multi-alias syntax](expand_multi_alias.html) (an example)
 
-  ## Inspect protocol
-
-  When inspecting a zipper, the default representation shows the current
-  node and provides indicators of the surrounding context without displaying
-  the full path, which can be very verbose for large ASTs.
-
-      iex> alias Sourceror.Zipper, as: Z
-      Sourceror.Zipper
-
-      iex> code = \"""
-      ...> def my_function do
-      ...>   :ok
-      ...> end\\
-      ...> \"""
-
-      iex> zipper = code |> Code.string_to_quoted!() |> Z.zip()
-      #Sourceror.Zipper<
-        #root
-        {:def, [line: 1], [{:my_function, [line: 1], nil}, [do: :ok]]}
-      >
-
-      iex> zipper |> Z.next()
-      #Sourceror.Zipper<
-        {:my_function, [line: 1], nil}
-        #...
-      >
-
-  This representation can be customized, however, using the `:custom_options`
-  field in `Inspect.Opts` and setting `:zipper` to one of the following:
-
-    * `:as_ast` (default) - display the current node as an AST
-    * `:as_code` - display the current node formatted as code
-    * `:raw` - display the raw `%Sourceror.Zipper{}` struct including the
-      `:path`.
-
-  Using the zipper defined above as an example:
-
-      iex> zipper |> inspect(custom_options: [zipper: :as_code]) |> IO.puts()
-      #Sourceror.Zipper<
-        #root
-        def my_function do
-          :ok
-        end
-      >
-
-      iex> zipper |> Z.next() |> inspect(custom_options: [zipper: :as_code]) |> IO.puts()
-      #Sourceror.Zipper<
-        my_function
-        #...
-      >
-
-      iex> zipper |> Z.next() |> inspect(custom_options: [zipper: :raw], pretty: true) |> IO.puts()
-      %Sourceror.Zipper{
-        node: {:my_function, [line: 1], nil},
-        path: %{
-          parent: %Sourceror.Zipper{
-            node: {:def, [line: 1], [{:my_function, [line: 1], nil}, [do: :ok]]},
-            path: nil
-          },
-          left: nil,
-          right: [[do: :ok]]
-        }
-      }
-
   """
 
   import Kernel, except: [node: 1]
@@ -565,76 +501,4 @@ defmodule Sourceror.Zipper do
 
   defp do_append_child(list, child) when is_list(list), do: list ++ [child]
   defp do_append_child({left, right}, child), do: {:{}, [], [left, right, child]}
-
-  defimpl Inspect do
-    import Inspect.Algebra
-
-    def inspect(zipper, opts) do
-      opts.custom_options
-      |> Keyword.get(:zipper, :as_ast)
-      |> inspect(zipper, opts)
-    end
-
-    defp inspect(:as_ast, zipper, opts) do
-      zipper.node
-      |> to_doc(opts)
-      |> inspect_opaque_zipper(zipper, opts)
-    end
-
-    defp inspect(:as_code, zipper, opts) do
-      zipper.node
-      |> Sourceror.to_algebra(Map.to_list(opts))
-      |> inspect_opaque_zipper(zipper, opts)
-    end
-
-    defp inspect(:raw, zipper, opts) do
-      open = color("%Sourceror.Zipper{", :map, opts)
-      sep = color(",", :map, opts)
-      close = color("}", :map, opts)
-      list = [node: zipper.node, path: zipper.path]
-      fun = fn kw, opts -> Inspect.List.keyword(kw, opts) end
-
-      container_doc(open, list, close, opts, fun, separator: sep)
-    end
-
-    defp inspect_opaque_zipper(inner, zipper, opts) do
-      opts = maybe_put_zipper_internal_color(opts)
-
-      inner_content =
-        [get_prefix(zipper, opts), inner, get_suffix(zipper, opts)]
-        |> concat()
-        |> nest(2)
-
-      force_unfit(
-        concat([
-          color("#Sourceror.Zipper<", :map, opts),
-          inner_content,
-          line(),
-          color(">", :map, opts)
-        ])
-      )
-    end
-
-    defp get_prefix(%Z{path: nil}, opts), do: concat([line(), internal("#root", opts), line()])
-
-    defp get_prefix(%Z{path: %{left: [_ | _]}}, opts),
-      do: concat([line(), internal("#...", opts), line()])
-
-    defp get_prefix(_, _), do: line()
-
-    defp get_suffix(%Z{path: %{right: [_ | _]}}, opts),
-      do: concat([line(), internal("#...", opts)])
-
-    defp get_suffix(_, _), do: empty()
-
-    defp internal(string, opts), do: color(string, :zipper_internal, opts)
-
-    # This prevents colorizing in contexts where no other syntax colors
-    # are present, e.g. in a test. Is there a better way to check this?
-    defp maybe_put_zipper_internal_color(%Inspect.Opts{syntax_colors: []} = opts), do: opts
-
-    defp maybe_put_zipper_internal_color(%Inspect.Opts{syntax_colors: colors} = opts) do
-      %{opts | syntax_colors: Keyword.put_new(colors, :zipper_internal, :light_black)}
-    end
-  end
 end
